move ConfigRepository to common module
