+++ common/build.gradle
@@ -70,6 +70,10 @@ dependencies {
 	implementation 'com.squareup.moshi:moshi:1.11.0'
 	kapt "com.squareup.moshi:moshi-kotlin-codegen:1.11.0"
 	implementation 'com.squareup.okhttp3:okhttp:4.7.2'
+	implementation 'com.squareup.retrofit2:retrofit:2.9.0'
+	implementation 'com.squareup.retrofit2:converter-moshi:2.9.0'
+	implementation 'com.squareup.okhttp3:logging-interceptor:4.7.2'
+	implementation 'androidx.security:security-crypto:1.0.0'
 
 	// QR code scanning
 	api "com.journeyapps:zxing-android-embedded:4.2.0"
