+++ verifier/src/main/java/ch/admin/bag/covidcertificate/verifier/HomeFragment.kt
@@ -22,9 +22,8 @@ import ch.admin.bag.covidcertificate.common.dialog.InfoDialogFragment
 import ch.admin.bag.covidcertificate.common.html.BuildInfo
 import ch.admin.bag.covidcertificate.common.html.HtmlFragment
 import ch.admin.bag.covidcertificate.common.util.AssetUtil
-import ch.admin.bag.covidcertificate.verifier.data.SecureStorage
+import ch.admin.bag.covidcertificate.common.data.ConfigSecureStorage
 import ch.admin.bag.covidcertificate.verifier.databinding.FragmentHomeBinding
-import ch.admin.bag.covidcertificate.common.faq.FaqFragment
 import ch.admin.bag.covidcertificate.verifier.faq.VerifierFaqFragment
 import ch.admin.bag.covidcertificate.verifier.pager.HomescreenPageAdapter
 import ch.admin.bag.covidcertificate.verifier.pager.HomescreenPagerFragment
@@ -113,7 +112,7 @@ class HomeFragment : Fragment() {
 			val hasInfoBox = localizedInfo != null
 
 			val onClickListener = localizedInfo?.let { infoBox ->
-				val secureStorage = SecureStorage.getInstance(notificationButton.context)
+				val secureStorage = ConfigSecureStorage.getInstance(notificationButton.context)
 				if (secureStorage.getLastShownInfoBoxId() != infoBox.infoId) {
 					closeCurrentInfoDialog()
 					showInfoDialog(infoBox)
