+++ wallet/src/main/java/ch/admin/bag/covidcertificate/wallet/data/WalletSecureStorage.kt
@@ -15,25 +15,15 @@ import android.content.SharedPreferences
 import androidx.security.crypto.EncryptedSharedPreferences
 import androidx.security.crypto.MasterKeys
 import ch.admin.bag.covidcertificate.eval.utils.SingletonHolder
-import ch.admin.bag.covidcertificate.common.config.ConfigModel
-import com.squareup.moshi.Moshi
 import java.io.IOException
 import java.security.GeneralSecurityException
 
-class SecureStorage private constructor(context: Context) {
+class WalletSecureStorage private constructor(context: Context) {
 
-	companion object : SingletonHolder<SecureStorage, Context>(::SecureStorage) {
+	companion object : SingletonHolder<WalletSecureStorage, Context>(::WalletSecureStorage) {
 
 		private const val PREFERENCES = "SecureStorage"
-
 		private const val KEY_ONBOARDING_COMPLETED = "onboarding_completed"
-		private const val KEY_CONFIG = "ConfigKey"
-		private const val KEY_CONFIG_LAST_SUCCESS = "LastSuccessTimestampKey"
-		private const val KEY_CONFIG_LAST_SUCCESS_APP_AND_OS_VERSION = "LastSuccessVersionKey"
-		private const val KEY_CONFIG_SHOWN_INFO_BOX_ID = "LastShownInfoBoxId"
-
-		private val moshi = Moshi.Builder().build()
-		private val configModelAdapter = moshi.adapter(ConfigModel::class.java)
 	}
 
 	private val prefs: SharedPreferences
@@ -75,25 +65,4 @@ class SecureStorage private constructor(context: Context) {
 
 	fun setOnboardingCompleted(completed: Boolean) = prefs.edit().putBoolean(KEY_ONBOARDING_COMPLETED, completed).apply()
 
-	fun updateConfigData(config: ConfigModel, timestamp: Long, appVersion: String) {
-		val editor = prefs.edit()
-		editor.putLong(KEY_CONFIG_LAST_SUCCESS, timestamp)
-		editor.putString(KEY_CONFIG_LAST_SUCCESS_APP_AND_OS_VERSION, appVersion)
-		editor.putString(KEY_CONFIG, configModelAdapter.toJson(config))
-		editor.apply()
-	}
-
-	fun getConfig(): ConfigModel? =
-		prefs.getString(KEY_CONFIG, null)
-			?.let { configModelAdapter.fromJson(it) }
-
-	fun getConfigLastSuccessTimestamp(): Long =
-		prefs.getLong(KEY_CONFIG_LAST_SUCCESS, 0)
-
-	fun getConfigLastSuccessAppAndOSVersion(): String? =
-		prefs.getString(KEY_CONFIG_LAST_SUCCESS_APP_AND_OS_VERSION, null)
-
-	fun setLastShownInfoBoxId(infoBoxId: Long) = prefs.edit().putLong(KEY_CONFIG_SHOWN_INFO_BOX_ID, infoBoxId).apply()
-
-	fun getLastShownInfoBoxId() : Long = prefs.getLong(KEY_CONFIG_SHOWN_INFO_BOX_ID, 0L)
 }
